+++ tests/sdk/llm/test_model_features.py
@@ -1,10 +1,7 @@
-from unittest.mock import patch
-
 import pytest
 
 from openhands.sdk.llm.utils.model_features import (
     get_features,
-    get_unverified_models,
     model_matches,
     normalize_model_name,
 )
@@ -268,37 +265,3 @@ def test_supports_stop_words_false_models(model):
     """Test models that don't support stop words."""
     features = get_features(model)
     assert features.supports_stop_words is False
-
-
-def test_organize_models_and_providers():
-    models = [
-        "openai/gpt-4o",
-        "anthropic/claude-sonnet-4-20250514",
-        "o3",
-        "o4-mini",
-        "devstral-small-2505",
-        "mistral/devstral-small-2505",
-        "anthropic.claude-3-5",  # Ignore dot separator for anthropic
-        "unknown-model",
-        "custom-provider/custom-model",
-        "openai/another-model",
-    ]
-
-    # monkeypatch.setattr(mf, "get_supported_llm_models", lambda *a, **kw: models)
-
-    with patch(
-        "openhands.sdk.llm.utils.model_features.get_supported_llm_models",
-        return_value=models,
-    ):
-        result = get_unverified_models()
-
-        assert "openai" in result
-        assert "anthropic" not in result  # don't include verified models
-        assert "mistral" not in result
-        assert "other" in result
-
-        assert len(result["openai"]) == 1
-        assert "another-model" in result["openai"]
-
-        assert len(result["other"]) == 1
-        assert "unknown-model" in result["other"]
